Update ResizeImage plugin.
