--- packages/ckeditor5-image/src/imageresize/imageresize.js
@@ -1,58 +1,0 @@
-/**
- * @license Copyright (c) 2003-2020, CKSource - Frederico Knabben. All rights reserved.
- * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
- */
-
-/**
- * @module image/imageresize
- */
-
-import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
-import ImageResizeUI from './ui/imageresizeui';
-import ImageResizeEditing from './imageresizeediting';
-
-import '../../theme/imageresize.css';
-
-/**
- * The image resize plugin.
- *
- * It adds a possibility to resize each image using handles.
- *
- * @extends module:core/plugin~Plugin
- */
-export default class ImageResize extends Plugin {
-	/**
-	 * @inheritDoc
-	 */
-	static get requires() {
-		return [ ImageResizeEditing, ImageResizeUI ];
-	}
-
-	/**
-	 * @inheritDoc
-	 */
-	static get pluginName() {
-		return 'ImageResize';
-	}
-}
-
-/**
- * The available options are `'px'` or `'%'`.
- *
- * Determines the size unit applied to the resized image.
- *
- *		ClassicEditor
- *			.create( editorElement, {
- *				image: {
- *					resizeUnit: 'px'
- *				}
- *			} )
- *			.then( ... )
- *			.catch( ... );
- *
- *
- * This option is used by the {@link module:image/imageresize~ImageResize} feature.
- *
- * @default '%'
- * @member {String} module:image/image~ImageConfig#resizeUnit
- */
